requirements.txt
Flask
yt-dlp
customtkinter==5.2.1
pillow==10.2.0 